refactor: 방 입장 로직 RoomEventService에서 RoomService로 이동
- RoomJoinEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RoomJoinConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 제거 및 코드 정리
