+++ backend/src/main/java/coffeeshout/room/application/service/RoomEventService.java
@@ -1,17 +1,11 @@
 package coffeeshout.room.application.service;
 
-import coffeeshout.room.domain.JoinCode;
-import coffeeshout.room.domain.Room;
 import coffeeshout.room.domain.RoomState;
-import coffeeshout.room.domain.event.RoomJoinEvent;
 import coffeeshout.room.domain.event.RouletteShowEvent;
 import coffeeshout.room.domain.event.RouletteShownEvent;
 import coffeeshout.room.domain.event.RouletteSpinEvent;
 import coffeeshout.room.domain.event.RouletteWinnerEvent;
-import coffeeshout.room.domain.player.PlayerName;
 import coffeeshout.room.domain.player.Winner;
-import coffeeshout.room.domain.service.RoomCommandService;
-import coffeeshout.room.infra.messaging.RoomEventWaitManager;
 import coffeeshout.room.infra.persistence.RoulettePersistenceService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,31 +17,9 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class RoomEventService {
 
-    private final RoomCommandService roomCommandService;
-    private final DelayedRoomRemovalService delayedRoomRemovalService;
     private final ApplicationEventPublisher eventPublisher;
-    private final RoomEventWaitManager roomEventWaitManager;
     private final RoulettePersistenceService roulettePersistenceService;
     private final RouletteService rouletteService;
-
-    public void joinRoom(RoomJoinEvent event) {
-        log.info("JoinCode[{}] 게스트 방 입장 이벤트 처리 - 게스트 이름: {}",
-                event.joinCode(),
-                event.guestName()
-        );
-
-        try {
-            final Room room = roomCommandService.joinGuest(
-                    new JoinCode(event.joinCode()),
-                    new PlayerName(event.guestName())
-            );
-
-            roomEventWaitManager.notifySuccess(event.eventId(), room);
-        } catch (Exception e) {
-            roomEventWaitManager.notifyFailure(event.eventId(), e);
-            throw e;
-        }
-    }
 
     public void showRoulette(RouletteShowEvent event) {
         log.info("JoinCode[{}] 룰렛 화면 표시 이벤트 처리", event.joinCode());
