+++ backend/src/main/java/coffeeshout/room/infra/messaging/consumer/RoomJoinConsumer.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.infra.messaging.consumer;
 
-import coffeeshout.room.application.service.RoomEventService;
+import coffeeshout.room.application.service.RoomService;
 import coffeeshout.room.domain.event.RoomJoinEvent;
 import java.util.function.Consumer;
 import lombok.RequiredArgsConstructor;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class RoomJoinConsumer implements Consumer<RoomJoinEvent> {
 
-    private final RoomEventService roomEventService;
+    private final RoomService roomService;
 
     @Override
     public void accept(RoomJoinEvent event) {
-        roomEventService.joinRoom(event);
+        roomService.joinRoom(event);
     }
 }
